More changes per review feedback.

+++ scripts/update_repos_version.py
@@ -51,7 +51,8 @@ class GitHubApiUtils(object):
 @click.command()
 @click.option("--module_name", help="Name of Python module which is being updated.", type=str, required=True)
 @click.option("--new_version", help="Updated version of Python module.", type=str, required=True)
-def bump_repos_version(module_name, new_version):
+@click.option("--local_only", help="Modify local repo branch without pushing to -or- creating PR on remote.", is_flag=True, default=False)
+def bump_repos_version(module_name, new_version, local_only):
     """
     Changes the pinned version number in the requirements files of all repos
     which have the specified Python module as a dependency.
@@ -59,7 +60,6 @@ def bump_repos_version(module_name, new_version):
     This script assumes that GITHUB_TOKEN is set for GitHub authentication.
     """
     # Make the cloning directory and change directories into it.
-    original_dir = os.getcwd()
     tmp_dir = tempfile.mkdtemp(dir=os.getcwd())
 
     # Iterate through each repository.
@@ -88,7 +88,7 @@ def bump_repos_version(module_name, new_version):
 
         # Search through all TXT files to find all lines with the module name, changing the pinned version.
         files_changed = False
-        for root, dirs, files in os.walk('.'):
+        for root, _dirs, files in os.walk('.'):
             for file in files:
                 if file.endswith('.txt') and (('requirements' in file) or ('requirements' in root)):
                     found = False
@@ -107,13 +107,17 @@ def bump_repos_version(module_name, new_version):
 
         if not files_changed:
             # Module name wasn't found in the requirements files.
-            logging.error("Module name '{}' not found in repo {} - skipping.".format(module_name, repo_url))
+            logging.info("Module name '{}' not found in repo {} - skipping.".format(module_name, repo_url))
             continue
 
         # Add/commit the files.
         ret_code = subprocess.call(['git', 'commit', '-am', 'Updating {} requirement to version {}'.format(module_name, new_version)])
         if ret_code:
             logging.error("Failed to add and commit changed files to repo {}".format(repo_url))
+            continue
+
+        if local_only:
+            # For local_only, don't push the branch to the remote and create the PR - leave all changes local for review.
             continue
 
         # Push the branch.
@@ -125,27 +129,28 @@ def bump_repos_version(module_name, new_version):
         # Create a PR with an automated message.
         rollback_branch_push = False
         try:
+            # The GitHub "mention" below does not work via the API - unfortunately...
             response = gh.create_pull(
                 title='Change {} version.'.format(module_name),
                 body='Change the required version of {} to {}.\n\n@edx-ops/pipeline-team Please review and tag appropriate parties.'.format(module_name, new_version),
                 head=branch_name,
                 base='master'
             )
-        except UnknownObjectException:
-            logging.error('No GitHub PR-creating permissions - did you set GITHUB_TOKEN?')
-            rollback_branch_push = True
         except:
-            logging.error('Failed to create PR for repo {}!'.format(repo_url))
+            logging.error('Failed to create PR for repo {} - did you set GITHUB_TOKEN?'.format(repo_url))
             rollback_branch_push = True
         else:
-            logging.info('Created PR for repo {}: {}'.format(repo_url, response.html_url))
+            logging.info('Created PR #{} for repo {}: {}'.format(response.number, repo_url, response.html_url))
 
         if rollback_branch_push:
+            # Since the PR creation failed, delete the branch in the remote repo as well.
             ret_code = subprocess.call(['git', 'push', 'origin', '--delete', branch_name])
             if ret_code:
                 logging.error("ROLLBACK: Failed to delete upstream branch {} for repo {}".format(branch_name, repo_url))
 
-    shutil.rmtree(tmp_dir)
+    if not local_only:
+        # Remove the temp directory containing all the cloned repos.
+        shutil.rmtree(tmp_dir)
 
 if __name__ == "__main__":
     bump_repos_version()
